refactor(Semibook): Cleanup
Consistent use of access modifiers, group related methods (including some from tradeEventManagement that had lost their way..)

Also fixes a bug in offerInfo: prev was not mapped from raw, so could be zero instead of undefined.
