+++ test/unit/tradeEventManagement.unit.test.ts
@@ -3,184 +3,17 @@ import assert from "assert";
 import { Big } from "big.js";
 import { BigNumber } from "ethers";
 import { describe, it } from "mocha";
-import {
-  deepEqual,
-  anything,
-  instance,
-  mock,
-  spy,
-  verify,
-  when,
-} from "ts-mockito";
-import { Semibook, Market, Token } from "../../src";
+import { anything, instance, mock, spy, verify, when } from "ts-mockito";
+import { Market, Token } from "../../src";
 import TradeEventManagement from "../../src/util/tradeEventManagement";
-import UnitCalculations from "../../src/util/unitCalculations";
 import {
   OfferFailEvent,
   OfferFailWithPosthookDataEvent,
   OfferSuccessEvent,
-  OrderCompleteEvent,
 } from "../../src/types/typechain/Mangrove";
 import { MangroveOrderStartEvent } from "../../src/types/typechain/MangroveOrder";
-import TickPriceHelper from "../../src/util/tickPriceHelper";
 
 describe("TradeEventManagement unit tests suite", () => {
-  describe("rawOfferToOffer", () => {
-    const rawGives = BigNumber.from(2);
-    const rawTick = BigNumber.from(1);
-
-    const rawOffer = {
-      id: BigNumber.from(1),
-      gasprice: BigNumber.from(2),
-      maker: "maker",
-      gasreq: BigNumber.from(0),
-      gives: rawGives,
-      tick: rawTick,
-    };
-
-    it("returns offer with correct values for bids", async function () {
-      //Arrange
-      const tradeEventManagement = new TradeEventManagement();
-
-      const marketSide: Market.BA = "bids";
-
-      const baseTokenMock = mock(Token);
-      const baseTokenDecimals: number = 3;
-      when(baseTokenMock.decimals).thenReturn(baseTokenDecimals);
-
-      const expectedGives = UnitCalculations.fromUnits(
-        rawGives,
-        baseTokenDecimals,
-      );
-      when(baseTokenMock.fromUnits(rawGives)).thenReturn(expectedGives);
-
-      const quoteTokenMock = mock(Token);
-      const quoteTokenDecimals = 1;
-
-      const tickPriceHelper = new TickPriceHelper(marketSide, {
-        base: { decimals: baseTokenDecimals },
-        quote: { decimals: quoteTokenDecimals },
-      });
-
-      const semibookMock = mock(Semibook);
-      when(semibookMock.tickPriceHelper).thenReturn(tickPriceHelper);
-      when(semibookMock.ba).thenReturn(marketSide);
-
-      const marketMock = mock(Market);
-      when(semibookMock.market).thenReturn(instance(marketMock));
-      when(marketMock.getOutboundInbound(marketSide)).thenReturn({
-        outbound_tkn: instance(baseTokenMock),
-        inbound_tkn: instance(quoteTokenMock),
-      });
-
-      const expectedPrice = tickPriceHelper.priceFromTick(rawTick);
-
-      // key difference between bids and asks here; for bids, we have volume = gives / price
-      const expectedVolume = expectedGives.div(expectedPrice);
-
-      // necessary to compare Big numbers with deepEqual in when() to have mock match expected values
-      when(
-        marketMock.getVolumeForGivesAndPrice(
-          marketSide,
-          deepEqual(expectedGives),
-          deepEqual(expectedPrice),
-        ),
-      ).thenReturn(expectedVolume);
-
-      //Act
-      const result = tradeEventManagement.rawOfferToOffer(
-        instance(semibookMock),
-        rawOffer,
-      );
-
-      //Assert
-      const expectedOffer: Market.OfferSlim = {
-        id: rawOffer.id.toNumber(),
-        gasprice: rawOffer.gasprice.toNumber(),
-        maker: rawOffer.maker,
-        gasreq: rawOffer.gasreq.toNumber(),
-        gives: expectedGives,
-        tick: rawOffer.tick,
-        price: expectedPrice,
-        wants: expectedPrice.mul(expectedGives).round(),
-        volume: expectedVolume,
-      };
-
-      assert.deepEqual(result, expectedOffer);
-    });
-
-    it("returns offer with correct values for asks", async function () {
-      //Arrange
-      const tradeEventManagement = new TradeEventManagement();
-
-      const marketSide: Market.BA = "asks";
-
-      const baseTokenMock = mock(Token);
-      const baseTokenDecimals: number = 3;
-
-      const quoteTokenMock = mock(Token);
-      const quoteTokenDecimals = 1;
-      const expectedGives = UnitCalculations.fromUnits(
-        rawGives,
-        quoteTokenDecimals,
-      );
-
-      when(quoteTokenMock.decimals).thenReturn(quoteTokenDecimals);
-      when(quoteTokenMock.fromUnits(rawGives)).thenReturn(expectedGives);
-
-      const tickPriceHelper = new TickPriceHelper(marketSide, {
-        base: { decimals: baseTokenDecimals },
-        quote: { decimals: quoteTokenDecimals },
-      });
-
-      const semibookMock = mock(Semibook);
-      when(semibookMock.tickPriceHelper).thenReturn(tickPriceHelper);
-      when(semibookMock.ba).thenReturn(marketSide);
-
-      const marketMock = mock(Market);
-      when(semibookMock.market).thenReturn(instance(marketMock));
-      when(marketMock.getOutboundInbound(marketSide)).thenReturn({
-        outbound_tkn: instance(quoteTokenMock),
-        inbound_tkn: instance(baseTokenMock),
-      });
-
-      const expectedPrice = tickPriceHelper.priceFromTick(rawTick);
-
-      // key difference between bids and asks here; for asks, we have volume = gives
-      const expectedVolume = expectedGives;
-
-      // necessary to compare Big numbers with deepEqual in when() to have mock match expected values
-      when(
-        marketMock.getVolumeForGivesAndPrice(
-          marketSide,
-          deepEqual(expectedGives),
-          deepEqual(expectedPrice),
-        ),
-      ).thenReturn(expectedVolume);
-
-      //Act
-      const result = tradeEventManagement.rawOfferToOffer(
-        instance(semibookMock),
-        rawOffer,
-      );
-
-      //Assert
-      const expectedOffer: Market.OfferSlim = {
-        id: rawOffer.id.toNumber(),
-        gasprice: rawOffer.gasprice.toNumber(),
-        maker: rawOffer.maker,
-        gasreq: rawOffer.gasreq.toNumber(),
-        gives: expectedGives,
-        wants: expectedPrice.mul(expectedGives).round(),
-        tick: rawOffer.tick,
-        price: expectedPrice,
-        volume: expectedVolume,
-      };
-
-      assert.deepEqual(result, expectedOffer);
-    });
-  });
-
   type summaryEvent = {
     args: {
       olKeyHash: string;
